add(소켓 <-> 타일 탈부착)
#1

+ 소켓에 타일 마운트 되면 색 변경
+ 타일 드래그 위치에 따라 바로 아래 소켓에 자동 부착 및 대상 소켓 마운트 true
