--- SlimeGG/Assets/Scripts/Commons/TileController.cs
@@ -1,17 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.EventSystems;
-
-public class TileController : MonoBehaviour
-{
-    float zCoor = 9f;
-
-    void OnMouseDrag()
-    {
-        Vector3 mousePos = new Vector3(Input.mousePosition.x, Input.mousePosition.y, zCoor);
-        Vector3 objPosition = Camera.main.ScreenToWorldPoint(mousePos);
-        objPosition.z = zCoor;
-        transform.position = objPosition;
-    }
-}
