+++ SlimeGG/Assets/Scripts/Maps/SocketController.cs
@@ -7,19 +7,16 @@ public class SocketController : MonoBehaviour
     SpriteRenderer socketBackSprite = null;
     Color colorTriggered = new Color(0.1603774f, 0.1603774f, 0.1603774f, 1f);
     Color colorNormal = new Color(0.2830189f, 0.2763439f, 0.2763439f, 1f);
+    public bool isMounted = false;
 
     void Start()
     {
         socketBackSprite = transform.Find("Socket Back").GetComponent<SpriteRenderer>();
     }
 
-    void OnMouseEnter()
+    private void Update()
     {
-        socketBackSprite.color = colorTriggered;
+        socketBackSprite.color = isMounted ? colorTriggered : colorNormal;
     }
 
-    void OnMouseExit()
-    {
-        socketBackSprite.color = colorNormal;
-    }
 }
